feat: change FilmSearchResultsFragment to FilmSearchResultFragment
