+++ app/src/main/java/com/northcoders/media_tracker_front/fragments/FilmSearchResultFragment.java
@@ -1,0 +1,62 @@
+package com.northcoders.media_tracker_front.fragments;
+
+import android.os.Bundle;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.SearchView;
+
+import com.northcoders.media_tracker_front.R;
+import com.northcoders.media_tracker_front.adapter.RecyclerViewInterface;
+import com.northcoders.media_tracker_front.databinding.FragmentFilmSearchResultsBinding;
+
+public class FilmSearchResultFragment extends Fragment implements RecyclerViewInterface {
+    private FragmentFilmSearchResultsBinding binding;
+
+    public FilmSearchResultFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+    }
+    @Override
+    public void onViewCreated(View view, @Nullable Bundle savedInstanceState){
+        super.onCreate(savedInstanceState);
+        Bundle bundle = this.getArguments();
+        String enteredQuery = bundle.getString("SearchQuery");
+        binding.SearchViewSearch.setQuery("SearchQuery",false);
+        binding.SearchViewSearch.clearFocus();
+
+        binding.SearchViewSearch.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
+            @Override
+            public boolean onQueryTextSubmit(String query) {
+                return false;
+            }
+
+            @Override
+            public boolean onQueryTextChange(String newText) {
+                return false;
+            }
+        });
+
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_film_search_results, container, false);
+    }
+
+    @Override
+    public void onItemClick(int position) {
+
+    }
+}
