--- app/src/main/java/com/northcoders/media_tracker_front/fragments/FilmSearchResultsFragment.java
@@ -1,31 +1,0 @@
-package com.northcoders.media_tracker_front.fragments;
-
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import com.northcoders.media_tracker_front.R;
-
-public class FilmSearchResultsFragment extends Fragment {
-
-    public FilmSearchResultsFragment() {
-        // Required empty public constructor
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_film_search_results, container, false);
-    }
-}
